Move Feed object to correct namespace
